refactor: util -> cli
