+++ pkg/cli/alert.go
@@ -1,4 +1,4 @@
-package util
+package cli
 
 import (
 	"bufio"
